+++ libs/markdown/module/render-markdown/render-markdown.component.ts
@@ -1,17 +1,28 @@
 import {
   Component,
+  ElementRef,
   Input,
   OnInit
 } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { MarkdownService } from '@gewd/markdown/service';
+import { MarkdownService, MarkdownCacheService } from '@gewd/markdown/service';
+import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 
 @Component({
   selector: 'gewd-markdown',
   templateUrl: './render-markdown.component.html',
-  styleUrls: ['./render-markdown.component.css']
+  styleUrls: ['./render-markdown.component.scss']
 })
 export class RenderMarkdownComponent implements OnInit {
+  @Input()
+  public showRawMd = false;
+
+  @Input()
+  public useCache = false;
+
+  @Input()
+  public allowMermaid = true;
+
   get markdown (): string {
     return this._markdown;
   }
@@ -20,17 +31,22 @@ export class RenderMarkdownComponent implements OnInit {
   set markdown (value: string) {
     if (this._markdown  !== value) {
       this._markdown = value;
-      this._htmlToShow$.next(value);
+      if (this.showRawMd) {
+        this._htmlToShow$.next(value);
+      }
       this.compile();
     }
   }
 
-  private _htmlToShow$ = new BehaviorSubject("");
+  private _htmlToShow$ = new BehaviorSubject<string|SafeHtml>("");
   private _markdown: string;
 
-  public parsedHtml$: Observable<string>;
+  public parsedHtml$: Observable<string|SafeHtml>;
 
-  constructor (private service: MarkdownService) {
+  constructor (private service: MarkdownService,
+               private element: ElementRef,
+               private _: DomSanitizer,
+               private cache: MarkdownCacheService) {
     this.parsedHtml$ = this._htmlToShow$;
   }
 
@@ -39,7 +55,25 @@ export class RenderMarkdownComponent implements OnInit {
   }
 
   private async compile() {
-    const parsedHtml = await this.service.compileMarkdown(this._markdown, true);
-    this._htmlToShow$.next(parsedHtml);
+    try {
+      if (this.useCache) {
+        const cachedMarkdown = await this.cache.getCached(this._markdown);
+        if (!!cachedMarkdown) {
+          this._htmlToShow$.next(this._.bypassSecurityTrustHtml(cachedMarkdown));
+          return;
+        }
+      }
+
+      const parsedHtml = await this.service.compileMarkdown(this._markdown, this.allowMermaid);
+
+      this._htmlToShow$.next(this._.bypassSecurityTrustHtml(parsedHtml));
+
+      if (this.useCache) {
+        this.cache.saveCached(this._markdown, parsedHtml);
+      }
+    } catch (er) {
+      console.error(er);
+      this._htmlToShow$.next(er);
+    }
   }
 }