+++ libs/markdown/service/markdown.service.ts
@@ -7,21 +7,30 @@ import {
   LoadMermaidInjectorToken,
   MarkdownWorker
 } from '@gewd/markdown/contracts';
+import { MarkdownCacheService } from './markdown-cache.service';
+
+
+const compiledRegex = /<div class="mermaid">([\s\S]*?)<\/div>/mg;
+const _mermaidNode = document.createElement('div');
+_mermaidNode.hidden = true;
+
+let renderId = 0;
 
 @Injectable({
   providedIn: 'root'
 })
 export class MarkdownService {
-
   private canTriggerMermaidLoad = false;
   private mermaidAddedToPage = false;
   private workerProxy: Remote<MarkdownWorker>;
+
   constructor (@Inject(LoadMarkdownWorkerInjectorToken)
                private readonly workerPayload: GetWorkerPayload,
                @Inject(LoadMermaidInjectorToken) @Optional()
-               private readonly loadMermaidUrl: string
+               private readonly loadMermaidUrl: string,
+               private readonly cache: MarkdownCacheService
   ) {
-    this.workerProxy = wrap<MarkdownWorker>( workerPayload.getWorker());
+    this.workerProxy = wrap<MarkdownWorker>(workerPayload.getWorker());
     if (workerPayload.options) {
       this.workerProxy.init(workerPayload.options);
     }
@@ -32,51 +41,85 @@ export class MarkdownService {
   }
 
   public async compileMarkdown (str: string, triggerMermaid = false): Promise<string> {
-    const parsedMarkdown = await this.workerProxy.compile(str);
+    let parsedMarkdown = await this.workerProxy.compile(str);
 
     if (triggerMermaid && parsedMarkdown.match(/class="mermaid"/)) {
-      this.triggerMermaidLoadScript();
+      await this.triggerMermaidLoadScript();
+
+      const mermaidInstance = (window as any).mermaid;
+      if (mermaidInstance) {
+        const matched = parsedMarkdown.match(compiledRegex);
+
+        for (const mermaid of matched) {
+          const innerContent = mermaid.replace(/&gt;/mg, '>')
+            .replace('<div class="mermaid">', '')
+            .replace('</div>', '');
+
+          let rendered = '';
+
+          const cached = await this.cache.getCachedPart('mermaid', innerContent);
+
+          if (!!cached) {
+            rendered = cached;
+          } else {
+            rendered = mermaidInstance.render(`sub${renderId++}`, innerContent,
+              () => { });
+            this.cache.saveCachedPart('mermaid', innerContent, rendered);
+          }
+
+          parsedMarkdown = parsedMarkdown.replace(mermaid, `<code class="mermaid">${rendered}</code>`);
+        }
+      }
     }
 
     return parsedMarkdown;
   }
 
   private triggerMermaidLoadScript () {
+    if (!this.canTriggerMermaidLoad) {
+      return Promise.resolve();
+    }
+
     if (this.canTriggerMermaidLoad && !this.mermaidAddedToPage) {
       this.mermaidAddedToPage = true;
+      document.body.appendChild(_mermaidNode);
 
-      const scriptTag = document.createElement('script');
-      scriptTag.src = this.loadMermaidUrl;
-      // trigger mermaid init
-      scriptTag.onload = () =>{
-        const config = {
-          startOnLoad:true,
-          flowchart:{
-            useMaxWidth:false,
-            htmlLabels:true
-          }
+      return new Promise((resolve, reject) => {
+
+        const scriptTag = document.createElement('script');
+        scriptTag.src = this.loadMermaidUrl;
+        // trigger mermaid init
+        scriptTag.onload = () => {
+          const config = {
+            // startOnLoad:true,
+            flowchart: {
+              useMaxWidth: true,
+              htmlLabels: true
+            }
+          };
+
+          const mermaid = (window as any).mermaid;
+          mermaid.initialize(config);
+
+          resolve();
+
+          mermaid.parseError = function(err) {
+            console.error('MarkdownService, Mermaid: ', err);
+          };
         };
 
-        const mermaid = (window as any).mermaid;
-        mermaid.initialize(config);
-        mermaid.init();
+        document.body.appendChild(scriptTag);
 
-
-        mermaid.parseError = function(err){
-          console.error("MarkdownService, Mermaid: ", err);
-        };
-      };
-
-      document.body.appendChild(scriptTag);
+      });
     }
 
     if (this.canTriggerMermaidLoad && this.mermaidAddedToPage) {
-      setTimeout(() => {
-        const mermaidInstance = (window as any).mermaid;
-        if (mermaidInstance) {
-          mermaidInstance.init();
-        }
-      }, 600);
+      const mermaidInstance = (window as any).mermaid;
+      if (mermaidInstance) {
+        return Promise.resolve();
+      }
+
+      return Promise.reject();
     }
   }
 }