+++ libs/markdown/service/public_api.ts
@@ -1,1 +1,2 @@
 export * from './markdown.service';
+export * from './markdown-cache.service';