+++ libs/lazy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.0] - 2020-02-25
+
+### [`gewd-lazy-module-component`]
+
+#### Added
+
+- `isLoading` can be also applied to a content-child so have it visible during load/building-time of the needed component
+- `[showError]="false"` if any of the error shouldn't be displayed
+
 ## [0.1.0]
 
 Start of Changelog